pyproject.toml
[tool.poetry]
name = "csm-rewards"
version = "0.1.0"
description = "CSM rewards"
authors = []
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
eth-wake = "^4.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
